Crate annealing

This crate provides an implementation of the simulated annealing algorithm for approximating the global minimum of a given function.

Choose the temperatures and the annealing schedule wisely: this is your way of controlling how long you will have to wait. Note that the minimum temperature must be reachable.

References:

Jason Brownlee, 2021, “Simulated Annealing From Scratch in Python”

Mykel J. Kochenderfer, Tim A. Wheeler, 2019, “Algorithms for Optimization”

Jonathan Woollett-Light, simple_optimization crate

Structs

SA

Simulated annealing

Enums

APF

Acceptance probability function

NeighbourMethod

Method of getting a random neighbour

Schedule

Annealing schedule

Status

Status function

Type Definitions

Bounds

Bounds of the parameter space

CustomStatus

Custom status function

Point

Point in the parameter space