Type Definition annealing::CustomStatus

pub type CustomStatus<'a, F, const N: usize> = Box<dyn FnMut(usize, F, F, [F; N], F, [F; N]) + 'a>;

Custom status function

It’s a Box’ed FnMut trait (see why here), which allows you not only to access values of the signature variables, but also to bring external variables and use them, too (for example, for storing results in an array).

See the print method for the signature explanation.